Resolved the `realpath` of the sandbox directory in tests.
This fixes test failures on systems (like macOS) where `/tmp` is a
symlink. One such scenario is a test which makes a file in the
sandbox, then makes a symlink to the file, and finally compares the
realpath of the symlink to the file. If a part of the path of the
original file in the sandbox is a symlink, this will cause the test to
fail. The fact that the sandbox path is not a fully resolved file name
is confusing. By instead getting the `realpath` of the sandbox during
setup, we can fix the current test failures, and avoid future ones.

Review: https://reviews.apache.org/r/66912/