Improved tests for resizing persistent volumes.
Now the `GrowVolume` and `ShrinkVolume` tests launch tasks after
resizing the volumes to ensure that the operations take effect on
agents. The `NonSpeculativeGrowAndLaunch` and
`NonSpeculativeShrinkAndLaunch` tests launch an additional task to
verify that the original volume consumed by the operations cannot be
used by subsequent tasks.

This patch also adjusted when the clock is resumed so schedulers will
not receive unexpected offers.

Review: https://reviews.apache.org/r/66920/