Failure to update registry should abort the master process.
When the registrar fails to update the registry it would abort the
actor and fail all future operations. However when the registrar
update is requested by an operator API such as a maintenance update,
the master process doesn't shut down (a 500 error is returned to the
client instead) and all subsequent operations will fail.

This patch fixes the specific maintenance API case but we can follow
up with other call sites or put a fix in for the registrar itself.

Review: https://reviews.apache.org/r/66919/