Fixed flakyness in 'MasterAPITest.MasterFailover'.
This test used to be sporadically segfault as described in MESOS-8687.
The suspected cause is that a in a master actor, the `httpSequence`
field was lazily initialized in `ProcessBase::consume()` and afterwards
a call to `ProcessBase::_consume()` was dispatched, where it was
assumed that `httpSequence` is already initialized.

However, during this test the master actor would be destroyed and a
new actor would be spawned with the same PID. The dispatched method
would be called on this new actor and find `httpSequence` to be not
initialized, leading to a crash.

This patch introduces a call to `Clock::settle()` after the master
is shut down to ensure the outstanding `_consume()` gets discarded
before starting the new master actor.

Review: https://reviews.apache.org/r/66799/