Bumped 'tox' version used by python to 3.0.0.
We require this version so that we can pass the '-qq' option to tox to
suppress its output. We want to see the output of the commands it
issues, but we don't want to see the output of tox itself. The '-qq'
option gives us this ability.

Reviewed at https://reviews.apache.org/r/64970/