Windows: Fixed quoted cases of `SubprocessTest.Flags` test.
Enabling the unit tests for flags with quotes in them required
switching from `cmd /c echo` to a native binary. The command prompt
built-in does not perform the same command line argument parsing that
native Windows binaries automatically do, so we added `test-echo.exe`
for the purposes of this test.

Note that the `BUILD_DIR` compile definition has to include the
configuration folder for the Visual Studio generator.

Review: https://reviews.apache.org/r/67075/