Windows: Implemented `os::shell` and enabled tests.
Once implemented, the tests had to be fixed to use Windows commands,
and a test that ensures `os::shell` doesn't hang was added. The
implementation does not reuse `os::spawn` because it must read all the
child process's piped output before waiting on it to exit, otherwise
it will deadlock.

Review: https://reviews.apache.org/r/67188