Updated cpplint to be compatible with Python 3.
This patch keeps the bundled cpplint at upstream `43d512b`.
Cpplint now works with Python 2 and 3 by adding part of the
content of https://github.com/google/styleguide/pull/349.

After aliasing `unicode` to `str`, and replacing `iteritems` and
`itervalues`, `cpplint.py` can work well with Python 3 if the linted
file does not have any issues. If there is even only one style
violation in the linted file, `cpplint.py` will need to an output
error message, then exceptions like `TypeError: write() argument must
be str, not bytes` will appear.

This is due to differing Unicode support between Python 2 and 3. In
Python 2 `str` is a byte sequence and `unicode` is used to handle
Unicode strings, but in Python 3 `str` is a Unicode sequence and
`unicode` does not exist. Type `bytes` is introduced to handle bytes
sequences, thus Python 3 `bytes` is equal to Python 2 `str`.

The type of the argument of `sys.stderr.write()` is kept the same
between Python 2 and 3, but the arguments of `sys.stderr.write()` are
changed. It receives a byte sequence in Python 2 and a Unicode
sequence in Python 3. As a result, `sys.stderr` should not be wrapped
by a UTF-8 stream reader and writer of codecs when using Python 3.

Review: https://reviews.apache.org/r/67055/