Ensured `SlaveRegisteredMessage`s trigger appropriate expectations.
An agent may retry `SlaveRegisteredMessage` if it does not receive
the registration confirmation on time. In this case the confirmation
may be sent twice. In tests with multiple registering agents, this
may result that an expectation set for one agent is satisfied by
a retried confirmation for another agent.

This patch unifies the way how this case is handled. An expectation
is augmented with a matcher for with the agent pid for which the
expectation is set.

Review: https://reviews.apache.org/r/67217