Windows: Fixed cURL copy command to run only once.
Previously this used `add_custom_target` on the `ExternalProject_Add`
target, `${CURL_TARGET}`. Because this is phony target, it caused the
copy operation to always run. By switching to
`ExternalProject_Add_Step`, we can change it to depend on the build
sub-step of the `ExternalProject_Add` logic, ensuring it runs
immediately after cURL is built, but not again.

Review: https://reviews.apache.org/r/67194/