Windows: Removed unnecessary dependencies from `windows/error.hpp`.
This function does not need to be Unicode, and so by using the ANSI
version instead, we can avoid the `stringify` (and `<codecvt>`)
dependency. We also only need to include `errorbase.hpp`, not
`error.hpp`.

Added unit tests for the `WindowsError` constructor. Removed the logic
to return early when the code is 0 because, on Windows, that still
corresponds to the message "The operation completed successfully."

Also formatted and stylized correctly.

Review: https://reviews.apache.org/r/67236