Disabled omission of frame pointers in debug builds.
This patch prevents the compiler from omitting frame pointers for
debug builds which can lead to a better debugging or profiling
experience.

Since having this flag turned on can disable certain optimizations we
do not enable it in non-debug builds. This can lead to minor
deviations in runtime behavior when combing release and debug flags
(e.g., when combining `--enable-debug` with `--enable-optimize`, or
with the cmake build type `RelWithDebInfo`), but still should give the
best possible build for pure release builds, and allow a reasonable
developer experience when debugging/profiling.

Review: https://reviews.apache.org/r/67108/