Windows: Fixed pipe inheritance in Mesos containerizer.
The Mesos containerizer needs an inheritable pipe that is not used for
stdio of the child process. It is used for signaling when the child
process should start running, so the pipe needs to be inheritable and
not overlapped.

Note this is going to change when handle white-listing is implemented,
as this control pipe will need to be passed explicitly, instead of the
current design where it is passed implicitly through an environment
variable.

Review: https://reviews.apache.org/r/66959/