Moved JWT test keys into the tests source file.
The JWT test keys were only used by the JWT tests, so there was
no necessity for them to be in a header file. Moved them into the
source file to fix `make distcheck` which was failing because the
`jwt_keys.hpp` header wasn't included in the distribution.

Review: https://reviews.apache.org/r/67285/