Fixed a performance issue in collect/await.
It turns out that even with C++11, some `std::list::size()`
implementations are still linear in complexity.

This was found when running perf against a metrics benchmark
for the master. A follow-up to avoid std::list in the collect.hpp
interfaces entirely is needed.

Review: https://reviews.apache.org/r/67298