Added python3 to list of Pylint excluded files.
This change ensures that pylint will not try to lint the new Python 3
support scripts if it is not run with Python 3. Having such a situation
results in unexpected errors such as "Unnecessary parens after 'print'
keyword". This change will not be applied in the Python 3 mesos-style.

Review: https://reviews.apache.org/r/67282/