Removed `destroyed` from `Container` struct in composing containerizer.
Previously, we stored `destroyed` promise for each container and used
it to guarantee that `destroy()` returns a non-empty value when the
destroy-in-progress stops an launch-in-progress using the next
containerizer. Since `wait()` and `destroy()` return the same
`ContainerTermination` value when called with the same ContainerID
argument, we can remove `destroyed` promise and add callbacks to
clean up `containers_` map instead.

Moreover, we added a clean up for terminated containers that have
been recovered after agent's restart.

Review: https://reviews.apache.org/r/66668/