Fixed python3/post-reviews.py type error.
The main challenge when switching from Python 2 to Python 3 is
to handle the strings, as many of them are now byte strings. I
have checked the other uses of `sys.stdout.buffer.write` while
working on this commit to check that it was the last issue of
that type in python3/post-reviews.py.

Review: https://reviews.apache.org/r/67321/