Modified launch.cpp to writes its logs into cerr.
`containerizer/mesos/launch.cpp` wrote some of its output into
`cout` and some into `cerr` without a clear distinction as to
why it was done in some cases.

Moreover, it is standard for logging utilities to write their
output into cerr, and finally in Mesos clog should probably
kept clean for the task output in cases there consumers of the
task which process the cout of it.

Review: https://reviews.apache.org/r/67339