Fixed python3/mesos-style.py type error.
The main challenge when switching from Python 2 to Python 3 is
to handle the strings, as many of them are now byte strings. I
have checked the other uses of `sys.stderr.write` while working
on this commit to check that it was the last issue of that type
in python3/mesos-style.py.

Review: https://reviews.apache.org/r/67340/