Enforced that `allowed_devices` entries must have a path.
If the operator omitted the `path` field from the `allowed_devices`
entry, the `cgroups/devices` isolator would silently ignore the
entry. Change this to an error to make the behavior consistent
with `linux/devices` and to make it clear to operators that the
configuration is incomplete.

Review: https://reviews.apache.org/r/67153/