Added bzip2, libarchive, and xz tarballs.
A subsequent commit will add libarchive to the build for all platforms
as well as modify the fetcher to use libarchive (instead of shelling
out to system binaries).

bzip2 and xz are being bundled for Windows only, as these utilities are
not commonly found on the Windows platform.
NOTE: The xz archive is modified from the original release to remove
files that are GNU GPL licensed, which are incompatible with the Apache
license.

Review: https://reviews.apache.org/r/67118/