Rejected non-zero scalar resource values which be represented as zero.
Internally values of scalar resources are stored in a fixed point
representation. This can lead to (expected) precision loss.

This patch updates the resource validation ensuring that values of
scalar resources are either zero or large enough so that they are
still represented as non-zero values after conversion to the internal
fixed point format. With that we can still parse resources containing
resource values explictly set to zero to be able to distinguish empty
and abset resources, but prevent propagation of values accidentially
rounded to an effectively empty result.

While this patch should not change valid, intended use cases, it might
change the behavior for erroneous, already broken framework workflows.

Review: https://reviews.apache.org/r/67274/