Fixed an overflow bug in Timeout that leads to hanging tests.
Due to the potential for overflow in `Timeout`
e.g. `after(Duration::max())`, a timer can be inserted into libprocess
with a negative time. This causes `Clock::settle()` with a paused
clock to get stuck because it thinks that a timer is overdue to run
but no tick is scheduled for it. A tick was scheduled, but it was for
a very large Duration.