Fixed an issue in metrics where timers accumulate.
Previously, we did not discard the `after(timeout)` future, which
results in a build-up of timers in libprocess (especially when
the timeout value is large). This updates the code to discard the
future if the metrics' futures complete.

It's still the case that we leave a very large clock tick scheduled
in libprocess for `Time::max()`, but the number of these scheduled
ticks should not grow beyond 1.