Introduced a random sorter as an alternative to the DRF sorter.
This sorter returns a weighted random shuffling of its clients
upon each `sort()` request.

This implementation is a copy of the `DRFSorter` with share
calculation logic (including the `dirty` bit) removed and an
adjusted `sort()` implementation. Work needs to be done to
reduce the amount of duplicated logic needed across sorter
implementations, but it is left unaddressed here.

Review: https://reviews.apache.org/r/67371/