Handled race condition when removing maintenance windows.
When executing the `Master::inverseOffers()` callback, it
could happen that the maintenance window the inverse offer
referred to was already removed by a concurrent call to
to the maintenance endpoint of Mesos.

In this case, we must not send out an inverse offer, because
having outstanding inverse offers for an agent without
any scheduled maintenance window will lead to a crash in
the allocator when attempting to remove this offer.

Review: https://reviews.apache.org/r/67403/