Windows: Fixed libarchive build arguments for bzip2 and xz.
The arguments for disabling bzip2 and xz were accidentally placed
in the Windows build rules (as well as the Posix rules, where they
should be).  This causes the libarchive to fallback to looking for
the associated executables (i.e. `bzip2 -d`) on Windows, which
usually fails because these are commonly not found on Windows.

This patch corrects linking rules to bzip2 and xz on Windows.