Fixed failing ArchiverTest due to long filename.
Replaced most of the ArchiverTest.ExtractZipFileWithLongDestinationDir
test such that it constructs a long path similar to how it is done in
filesystem_tests.cpp. This resolves a failure in the Jenkins BuildBot
job that runs all of the tests inside a Docker container, which uses a
filesystem called AUFS, which has an individual file name limit of 242
characters.

Review: https://reviews.apache.org/r/67454/