Windows: Log a fatal error if `WindowsFD(int)` is incorrectly used.
This constructor _should_ be `explicit` as the other constructors are;
however, for historical reasons it is implicit in order to support the
semantics of assigning from `-1, 0, 1, 2`. Unfortunately, its
implicitness leads to subtle bugs. For instance, the code `int_fd s =
::socket(...)` on Windows compiles, but behaves incorrectly because
the RHS `SOCKET` type is implicitly converted to an `int` to satisfy
the only available implicit constructor, since the `WindowsFD(SOCKET)`
constructor is marked `explicit`. As we should never construct off any
of these values, but cannot constrain it at compile-time, we should
add a fatal runtime error instead.

Review: https://reviews.apache.org/r/67465/