Moved some constant definitions into header file.
Since we are using C++11 and the declared constants are either POD or
`constexpr` types, we can define them in the header as `constexpr`.
This also removes the need to declare the constants as `extern` which
should make using their headers easier (e.g., simpler linkage
requirements).

Review: https://reviews.apache.org/r/67522/