Fixed the flakiness in the `NVIDIA_GPU_NvidiaDockerImage` test.
This test is flaky because it tries to download the 1GB 'nvidia/cuda'
image from Docker Hub, which might take more than 1 minute and cause
the command executor unable to register in time.

This patch fixes this problem by using the default executor, which does
not wait for fetching task images before registration. If the image
fetch stalls more than 1 minute, the container will fail because of the
`--fetcher_stall_timeout` agent flag.

The time we wait for `TASK_FINISHED` is also extended to 180 seconds.

Review: https://reviews.apache.org/r/67596