Fixed break of `port_mapping_tests.cpp`.
The last change applied to the file descriptor whitelist chain removed
the default value from the new argument, and this function was not
updated. The other change made was that the three FD arguments were
turned into one `ContainerIO`. This was missed because this entire
isolator and its associated tests are not in the CMake build.

Review: https://reviews.apache.org/r/67600