Windows: Fixed libarchive and zlib linkage warnings.
Our build of libarchive was erroneously linking in the shared version
of our build of zlib, and so defined `ZLIB_DLL`, which caused warnings
when the consumers of libarchive also linked in the static version of
zlib. Instead, we can direct libarchive to use exactly the zlib
library the rest of the build is using.

Also forwarded the C arguments to libarchive for consistency with the
rest of our dependencies.

Review: https://reviews.apache.org/r/67607/