Made `PerfEventSubsystemProcess` can be created without `--perf_events`.
If the agent flag `--perf_events` is not specified, the `perf_event`
subsystem (i.e., `PerfEventSubsystemProcess`) can still be created, but
it will not do sampling at all. The reason that we want to do this is,
in some cases, even it is not required to sample any events, we still
want each container to have its own `perf_event` cgroup instead under
the root cgroup.

Review: https://reviews.apache.org/r/67574/