Updated ZooKeeper retry logic to retry on `ENOENT` too.
Per MESOS-3790, the call to `zookeeper_init` maps `EAI_NONAME` and
`EAI_NODATA` to an `errno` value of `ENOENT`, and all others except
`EAI_MEMORY` to `EINVAL`. Mesos's ZooKeeper logic is written to retry
this initialization for ten minutes if the error is `EINVAL`, and
should be updated to also retry if the error is `ENOENT`.

This is necessary because if the initialization is not retried, the
process crashes due to the `PLOG(FATAL)` call, and if it crashes, it
will interrupt other Mesos threads and potentially leave the
environment in an unknown state. For instance, we have seen
intermittent failures where the systemd unit file
`mesos_executors.slice` is created but empty because Mesos crashed
between creating the file and flushing the write to the file. This
then leads to errors when the agent is restarted (and succeeds to
connect to ZooKeeper), because the agent explicitly does not attempt
to write to the unit file if it already exists.

Review: https://reviews.apache.org/r/67587/