Fixed archiver failure to extract hardlinks to another directory.
The archiver utility may fail to extract hardlinks when the destination
target is different than the current working directory.  This happens
because hardlinks are generally expressed in relative terms within
an archive, but the target should be changed to point to the extracted
location during extraction.

This updates the archiver utility to handle links correctly and adds
a regression test.

Review: https://reviews.apache.org/r/67680