Always defined `PICOJSON_USE_INT64`.
The macro `PICOJSON_USE_INT64` must always be defined when attempting
to use stout's JSON facilities since they unconditionally depend on
these features.

Previously, we relied on the Mesos build system to set that macro from
the command line, but since we don't generate a separate pkg-config
file for stout there is no way to convey this information to other
users of stout.

Even for users trying to use stout as part of a libmesos installation,
it might be surprising that it is required to read `CPPFLAGS` if only
the header-only stout part is to be used.

Review: https://reviews.apache.org/r/67632/