Added `io::prepare_async` and `io::is_async` functions.
The `os::nonblock` and `os::isNonBlock` fucntions are stubs that only
do things for Windows sockets. For the IOCP implementation, we need to
associate the handles with the IOCP handel, which is created by
libprocess. Thus, we need these new functions to prepare the handle
for the async IO functions by either setting the hande to non-blocking
on POSIX systems or by associating it with an IOCP handle on Windows.

Review: https://reviews.apache.org/r/67386/