Added CMake `ENABLE_LIBWINIO` option for Windows IOCP backend.
This option will let us enable the new Windows IOCP backend for
libprocess. It can only be used on Windows, and it is mutually
exclusive with `ENABLE_LIBEVENT`. For now, it is disabled by default,
but this is expected to change, followed be a removal of libevent
support for Windows.

Review: https://reviews.apache.org/r/67391/