Windows: Wrapped docker executor code in job object.
One of the docker tests revealed that the docker executor code could
leak some of the docker processes if it gets killed. To avoid this, we
run the docker executor and all processes launched by it in a job
object, so that `os::killtree` will properly kill the process tree.

Review: https://reviews.apache.org/r/67456/