Removed incorrect symlink handling in archiver utility.
While fixing MESOS-9008, we accidentally assumed that all links needed
to be corrected when extracting to a different directory.  This is
necessary for hardlinks because a hardlink must point to a file
extracted earlier in the archive.  However, for symlinks, this is not
the case.  When extracted with no header modifications, symlinks will
point to their intended target (i.e. relative symlinks will not be
transformed into targetting absolute paths).

Review: https://reviews.apache.org/r/67780/