Updated configure.ac and Makefile.am to use `$PYTHON` not `python`.
This ensures that we use Python 2 even if the default Python
installed in the path under `python` is Python 3.

Previously, even when explicitly setting PYTHON_VERSION or PYTHON to a
version of Python 2.6+, configuration would succeed, but compilation
would fail. This commit fixes that.

Review: https://reviews.apache.org/r/67487/