Prevented resource providers from changing their name or type.
Since the agent uses e.g., a resource provider's name or type to
construct paths to persist resource provider state, changes to this
information on resource provider resubscription are not supported.

This patch persists a resource provider's name and type in the
resource provider registry and rejects a resource provider
resubscription if incompatible changes are detected. Since we did not
persist this information previous to mesos-1.7.0 we cannot and do not
perform validation against resource provider registry information
stored with earlier versions of Mesos.

Review: https://reviews.apache.org/r/67671/