Refactored base of Python CLI tests.
This patch mostly adds additional hardening to make sure we error out
with proper error messages in places where we weren't before.

However, as part of this, we also fixed a bug related to referencing
counting Master, Agent, and Task objects. Previously, this reference
count would increased in '__init__()' and decreased in '__del__()'.
However, this caused problems when limiting the number of agents and
masters to exactly 1, because we could never guarantee when the python
garbage collector would kick in to delete objects from previous test
runs. This manifested itself as flaky test failures. To fix this, we
now reference count only on explicit counts to 'launch()' and 'kill()'
instead of in '__init__()' and '__del__()'.

Review: https://reviews.apache.org/r/67842/