Allowed resources to be removed when updating the sorter.
This patch allows the allocator to remove resources when calling
`Sorter::update`, so Mesos can remove resources that are no longer
available through resource conversions.

An example usage is that we can make `DESTROY_VOLUME` to convert a
volume with a stale profile to an empty resource, so the disk space
will disappear instead of being offered out to other frameworks.

Review: https://reviews.apache.org/r/67669