Allowed resources to be removed in the hierarchical allocator.
When applying `DESTROY_VOLUME` or `DESTROY_BLOCK` on a CSI volume with a
stale profile, the volume will be converted to an empty resource to
avoid it to be offered to frameworks. This patch updates the
hierarchical allocator to allow such behaviors in `updateAllocation`.

Review: https://reviews.apache.org/r/67876