Fixed the error log in POSIX `os::rmdir`.
The POSIX implementation of `os::rmdir` uses FTS functions to walk
through all paths under a given directory. According to the manpage,
the `fts_path` field "contains the path specified to fts_open() as a
prefix," so should not be joined with the given directory again. See:
http://man7.org/linux/man-pages/man3/fts.3.html

Review: https://reviews.apache.org/r/67881