Fixed `build-virtualenv` to correctly clear virtual environment.
Running `build-virtualenv` with Python 3 followed by running it again
with Python 2 leaves the directory with both versions of Python in the
virtualenv. This allows Python 3 to still be used from within the
virtualenv, which is incorrect.

This is caused by a long outstanding issue in `virtualenv` that we now
handle by unconditionally removing the 'support/.virtualenv' directory
before building it again.

Review: https://reviews.apache.org/r/67911/