Fixed linting errors that appear with the recently updated pylint.
As part of this, we needed to remove some unnecessary delegate to
'super' calls that are legal in python2 but not in python3. That said,
they are also unnecessary in python2 so we remove them in both cases.

Specifically, we previously used to override 'main()' in our Python and
JS linters with:

    def main(self, modified_files):
        super(PyLinter, self).main(modified_files)

    def main(self, modified_files):
        super(JsLinter, self).main(modified_files)

However, in python3 this gives us linting errors because such a
delegation is unecessary. Just using the 'main()' function from the base
class is sufficient.

Review: http://reviews.apache.org/r/67910/