Improved performance of cgroups::read by verifying after failure.
It turns out that cgroups::verify is expensive and leads to a severe
performance issue on the agent during container metrics collection
if there are a lot of containers on the agent. See MESOS-8418.

Since cgroups::verify serves to provide a helpful error message,
this patch preserves the error message, but only if the read fails.

Longer term, there probably needs to be some re-structuring of the
code to make verification caller-controlled, or perhaps the verify
code can occur consistently post-operation (as done in this patch),
or perhaps verify can be optimized substantially.

Review: https://reviews.apache.org/r/67923/