Windows: Added `nullptr` checks when using `libwinio_loop` pointer.
It was discovered that the `Socket` constructor could dereference a
null pointer (by way of `prepare_async()`) if the Windows IOCP event
loop had not yet been initialized. So now we check for its
initialization before each dereference, and return an error or fatal
log event.

In order to ensure that it is initialized in `test-linkee`, we call
`process::initialize()`. This should be fixed in the future, per
MESOS-9097.

Review: https://reviews.apache.org/r/67976