Windows: Enabled `RemoteLink` tests.
This resolves MESOS-5941. Note that on Windows, `os::killtree()` is
only valid for processes specifically spawned in a group via a job
object, otherwise it will fail (and the failure is not being checked
here). Since the `linkee` process only ever spawns the single OS
process `test-linkee`, it is safe to switch to `os::kill()` which can
correctly kill it.

Review: https://reviews.apache.org/r/67978