Fixed issues with the JSON serialization tests.
There were some issues in the existing JSON serialization tests
when it comes to rapidjson. For example:

  - 0x7F does not need to be escaped per ECMA-404. This is what
    rapidjson adheres to.
  - Forward slash does not need to be escaped.

This updates the test to have a valid UTF-8 case with a multi-byte
sequence (which does not need to be escaped per ECMA-404), as well
as an invalid case where a multibyte sequence is partial (which we
allow for now since we do not have validation errors exposed).

Review: https://reviews.apache.org/r/67989