Avoid resource copying while serving state json.
The state serving code was constructing `Resources` wrappers in
order to jsonify resources. This incurs a copy of the underlying
Resource objects which is unnecessary and can be costly. This
removes the copy by proving a direct way to jsonify
`RepeatedPtrField<Resource>`.

This shows a small 10-15% reduction in the master StateQuery
benchmark timings:

                       min    q1    q3   max
baseline              3.48  3.54  4.12  4.40
after this patch      2.82  3.11  3.31  3.71

Review: https://reviews.apache.org/r/67993