Shared GC actor between Agent and MesosContainerizer.
This inserts the pointer to the GC actor into the MesosContainerizer,
so the containerizer can schedule sandboxes it manages for GC according
to the agent's GC policy.  The Composing/Docker Containerizers are
unchanged as they do not directly deal with nested or standalone
containers.

To minimize test changes, the GC actor argument defaults to nullptr.
Only tests that exercise sandbox GC will need to supply the actor.

Review: https://reviews.apache.org/r/68066