UI: Pull up the leader URL generation to a top-level function.
This path pulls up the leader URL prefix generation to a top-level
function in order to have it re-used across controllers.

This also ensures that everything shown is consistently coming
from the leading master, specifically, now the log viewer shows
the leading master's log (and the UI messaging is updated in this
patch accordingly).

Having this logic in just one place makes it easier to modify it,
e.g., to make it possible to use the UI via a reverse proxy.

Review: https://reviews.apache.org/r/68120