Refactored some cgroups helpers to do verify from callers.
Prior to this patch, some of the cgroups helpers call `cgroups::verify`
to make sure the hierarchy and cgroup are valid. This causes some
performance issues for the agent because `cgroups::verify` will read the
entire mount table, and mount table read is expensive if the mount table
is large. See more details in MESOS-8418.

In most of the cases, the verification has already been done when those
cgroups helpers are called. Therefore, it is better to let the caller do
the verification and make those cgroups helpers pure helpers.

Also, the verification is subject to race anyway. The cgroups global
state might change (e.g., operators modify the cgroups) after the
verification is done, making the verification itself not that super
useful.

This patch refactored the cgroups helpers and moved the verifications to
the callers.

Review: https://reviews.apache.org/r/68426