Replaced exchange in MpscLinkedQueue::dequeue with load/store.
There's only a single consumer, so we don't need to exchange
in dequeue. Testing shows that a load/store pair is much
more efficient, as it needs fewer CPU cycles. Especially
in the empty queue case, where the store does not have to
be executed at all.

Review: https://reviews.apache.org/r/68149/