Fixed an authentication request amplification issue in the master.
Per MESOS-9144, re-enqueuing authentication requests leads to an
amplification effect which can overwhelm the master if requests
continue to arrive rapidly on a heavily loaded master. This patch
avoids the re-enqueing and ensures the master immediately processes
a new authentication request for a client.

Review: https://reviews.apache.org/r/68306