Added methods to remove resource providers from provider manager.
This patch adds a method to remove a resource provider from the
resource provider manager. The resource provider will be marked as
removed in the manager's registry and disconnected. We also expose a
new `REMOVE` event whenever a resource provider was removed.

This patch does not add integration with e.g., the agent.

Review: https://reviews.apache.org/r/68144/