Sent an event to resource providers when they are removed.
In order to allow proper cleanup the resource provider manager sends a
`REMOVED` to a resource provider when it is being removed.

The event is not sent reliably, i.e., if the resource provider was
e.g., not subscribed when it was removed we currently will never
attempt to resend the event.

Review: https://reviews.apache.org/r/68145/