Updated libprocess filtering to take the Process UPID.
When `EXPECT_DISPATCH` was originally introduced in de8e10830d178a7c,
the pid was added to `DispatchEvent` rather than being passed through
the filter API. This adds a performance overhead (we must perform an
unnecessary `UPID` copy construction / destruction for every
dispatch), and it also requires that we inject the `UPID` into
additional events (for example, if we wanted to introduce
`FUTURE_EXITED`, we need to add the UPID into `ExitedEvent`).

Rather than continue this trend, we can instead pass the Process
`UPID` through the filter API so that the client knows which Process
the event belongs to. This also lets us remove the `UPID` copy
construction / destruction cost from `DispatchEvent`, which should
provide a minor performance benefit.

Review: https://reviews.apache.org/r/68325