Windows: Made `libwinio` the default option for libprocess.
Note that we now add `io_tests.cpp` unconditionally, which will
fail/hang with libevent. However, this is because libevent is buggy,
and should not be used. We would prefer the tests not pass with
libevent enabled, rather than appear to pass because the failing tests
are excluded, giving the false impression that it works correctly.

Review: https://reviews.apache.org/r/68329/