Added /files API test for queries with encoded sequences.
Due to a bug in the HTTP client in libprocess (MESOS-9168), the use of
percent-encoded sequences in paths did not work when using the HTTP
client in libprocess. This adds a test to ensure that the /files API
correctly handles these sequences in file paths.

Review: https://reviews.apache.org/r/68420