Fixed an expectation in the master tests.
In `MasterTest.TaskStateMetrics`, since the clock is resumed at the
end of the test, it's possible that the allocation interval may elapse
on a heavily-loaded machine and the scheduler may receive additional
offers.

This patch updates an expectation to accommodate additional offers
sent to the scheduler.

Review: https://reviews.apache.org/r/68345/