Added `ROOT_CGROUPS_LaunchNestedDebugAfterUnshareMntNamespace` test.
This test verifies detection of task's `mnt` namespace for a debug
nested container. Debug nested container must enter `mnt` namespace
of the task, so the agent tries to detect task's `mnt` namespace.
This test launches a long-running task which runs a subtask that
unshares `mnt` namespace. The structure of the resulting process tree
is similar to the process tree of the command executor (the task of
the command executor unshares `mnt` ns):

  0. root (aka "nanny"/"launcher" process) [root `mnt` namespace]
    1. task: sleep 1000 [root `mnt` namespace]
      2. subtaks: sleep 1000 [subtask's `mnt` namespace]

We expect that the agent detects task's `mnt` namespace.

Review: https://reviews.apache.org/r/68408/