Added const versions of slave-counting functions in Master.
The family of functions '_slaves_active()', etc. is logically
const, although the function signatures do not currently
reflect this due to MESOS-4995.

Since a const version of this functionality is required for
the changes introduced in the subsequent commits, it is
added with different names as a workaround.

Review: https://reviews.apache.org/r/68342/